Fix DefectModel/OperatorModel session issue and remove flask-sqlalchemy dependency
Removing the flask-sqlalchemy dependency has been on the todo list since fsa was needed in the popup.
The operator model was built without it, but then it was not using the same session as the defect.
This caused some issues (invalidated transaction errors - couldn't save defects),
the best way to fix them was to finally remove fsa.